feat: find venues

+++ src/main/java/com/example/BookEatNepal/Service/VenueService.java
@@ -12,7 +12,7 @@ public interface VenueService {
     String delete(int id);
 
 
-    VenueDTO findAll(String name, String location, int rating, int page, int size);
+    VenueDTO findAll(String name, String location, double rating, int page, int size);
 
     Venue findById(int id);
